editoast: move infra_cache tests from views to infra_cache
