+++ editoast/src/views/infra/mod.rs
@@ -1694,62 +1694,4 @@ pub mod tests {
             .await
             .assert_status_forbidden();
     }
-
-    // TODO: move it to the infra_cache crate
-    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-    async fn load_infra_cache() {
-        let app = test_app!().build();
-        let db_pool = app.db_pool();
-        let infra = create_empty_infra(&mut db_pool.get_ok()).await;
-        let infra_caches = dashmap::DashMap::new();
-        InfraCache::get_or_load(
-            &mut db_pool.get_ok(),
-            &infra_caches,
-            &infra,
-            &app.valkey_client(),
-            app.config().app_version.as_deref(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(infra_caches.len(), 1);
-        InfraCache::get_or_load(
-            &mut db_pool.get_ok(),
-            &infra_caches,
-            &infra,
-            &app.valkey_client(),
-            app.config().app_version.as_deref(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(infra_caches.len(), 1);
-    }
-
-    // TODO: move it to the infra_cache crate
-    #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-    async fn load_infra_cache_mut() {
-        let app = test_app!().build();
-        let db_pool = DbConnectionPoolV2::for_tests();
-        let infra = create_empty_infra(&mut db_pool.get_ok()).await;
-        let infra_caches = dashmap::DashMap::new();
-        InfraCache::get_or_load_mut(
-            &mut db_pool.get_ok(),
-            &infra_caches,
-            &infra,
-            &app.valkey_client(),
-            app.config().app_version.as_deref(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(infra_caches.len(), 1);
-        InfraCache::get_or_load_mut(
-            &mut db_pool.get_ok(),
-            &infra_caches,
-            &infra,
-            &app.valkey_client(),
-            app.config().app_version.as_deref(),
-        )
-        .await
-        .unwrap();
-        assert_eq!(infra_caches.len(), 1);
-    }
 }
